Move ImageFormat extension methods to Services.Imaging namespace
