+++ Mosey/Services/ScanningDevice.cs
@@ -653,37 +653,4 @@ namespace Mosey.Services
             return MemberwiseClone();
         }
     }
-
-    public static class ImageFormatExtensions
-    {
-        /// <summary>
-        /// Convert a file extension string to an <see cref="ScanningDevice.ImageFormat"/> <see cref="Enum"/>.
-        /// </summary>
-        /// <param name="value">An <see cref="ScanningDevice.ImageFormat"/> instance</param>
-        /// <param name="imageFormatStr">A file extension <see cref="string"/></param>
-        /// <returns>An <see cref="ScanningDevice.ImageFormat"/> <see cref="Enum"/></returns>
-        public static ScanningDevice.ImageFormat FromString(this ScanningDevice.ImageFormat value, string imageFormatStr)
-        {
-            if (Enum.TryParse(imageFormatStr, ignoreCase: true, out value))
-            {
-                return value;
-            }
-            else
-            {
-                throw new ArgumentException($"{imageFormatStr} is not a supported image file extension.");
-            }
-        }
-
-        /// <summary>
-        /// Convert to a <see cref="System.Drawing.Imaging.ImageFormat"/> instance.
-        /// </summary>
-        /// <param name="value"></param>
-        /// <returns>A <see cref="System.Drawing.Imaging.ImageFormat"/> instance</returns>
-        public static System.Drawing.Imaging.ImageFormat ToDrawingImageFormat(this ScanningDevice.ImageFormat value)
-        {
-            return (System.Drawing.Imaging.ImageFormat)typeof(System.Drawing.Imaging.ImageFormat)
-                    .GetProperty(value.ToString(), System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.IgnoreCase)
-                    .GetValue(null);
-        }
-    }
 }
